Reverted the change made to `loader` and `action` functions where the `accessToken` is non-null asserted in TypeScript, it now handles situations where `accessToken` isn't available yet.

+++ client/src/routes/note-delete.tsx
@@ -9,13 +9,14 @@ import {
 
 interface DeleteNoteActionArgs extends ActionFunctionArgs {
   params: Params<ParamParseKey<typeof Paths.noteDetail>>;
-  accessToken: string;
+  accessToken: string | null;
 }
 
 export async function deleteNoteAction({
   accessToken,
   params,
 }: DeleteNoteActionArgs) {
+  if (!accessToken) throw Error('You are not authenticated.');
   await deleteNote({
     accessToken,
     id: parseInt(params.noteId!),
